+++ tests/test_no_internal_clock/main.lua
@@ -15,7 +15,7 @@ local count3 = dut.count3:chdl()
 local count4 = dut.count4:chdl()
 
 local is_verilator = cfg.simulator == "verilator"
-local is_inertial_put = os.getenv("CFG_USE_INERTIAL_PUT") == "1"
+local is_inertial_put = os.getenv("VL_USE_INERTIAL_PUT") == "1"
 
 ---@param clk verilua.handles.CallableHDL
 ---@return verilua.scheduler.TaskFunction